风控测试后台/tutorial/requirements.txt
asn1crypto==0.24.0
boto3==1.9.225
botocore==1.12.225
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
DBUtils==1.3
Django==2.2.5
djangorestframework==3.10.3
docutils==0.15.2
idna==2.8
jmespath==0.9.4
pycparser==2.19
Pygments==2.4.2
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
